requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
anyio=2.0.2=py39h6e9494a_4
appdirs=1.4.4=pyh9f0ad1d_0
appnope=0.1.2=py39h6e9494a_1
argon2-cffi=20.1.0=py39h5a22ff9_2
async_generator=1.10=py_0
attrs=20.3.0=pyhd3deb0d_0
babel=2.9.0=pyhd3deb0d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
black=20.8b1=py_1
bleach=3.2.1=pyh9f0ad1d_0
brotlipy=0.7.0=py39h66d5b7b_1001
ca-certificates=2020.12.5=h033912b_0
certifi=2020.12.5=py39h6e9494a_1
cffi=1.14.4=py39h7786acb_1
chardet=4.0.0=py39h6e9494a_1
click=7.1.2=pyh9f0ad1d_0
cryptography=3.3.1=py39h79a2c39_1
cycler=0.10.0=py_2
dataclasses=0.7=pyhb2cacf7_7
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=pyhd8ed1ab_1003
freetype=2.10.4=h3f75d11_0
icu=68.1=h74dc148_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=3.4.0=py39h6e9494a_0
importlib_metadata=3.4.0=hd8ed1ab_0
ipykernel=5.4.2=py39h71a6800_0
ipython=7.19.0=py39h71a6800_0
ipython_genutils=0.2.0=py_1
isort=5.7.0=pyhd8ed1ab_0
jedi=0.18.0=py39h6e9494a_2
jinja2=2.11.2=pyh9f0ad1d_0
joblib=1.0.0=pyhd8ed1ab_0
jpeg=9d=hbcb3906_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.11=pyhd8ed1ab_1
jupyter_core=4.7.0=py39h6e9494a_0
jupyter_server=1.2.1=py39h6e9494a_0
jupyterlab=3.0.3=pyhd8ed1ab_0
jupyterlab_code_formatter=1.4.2=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.1.1=pyhd8ed1ab_0
kiwisolver=1.3.1=py39hedf5dff_1
lcms2=2.11=h11f7e16_1
libblas=3.9.0=7_openblas
libcblas=3.9.0=7_openblas
libcxx=11.0.0=h4c3b8ed_1
libffi=3.3=h046ec9c_2
libgfortran=5.0.0=h6c81a4c_15
libgfortran5=9.3.0=h6c81a4c_15
liblapack=3.9.0=7_openblas
libopenblas=0.3.12=openmp_h54245bb_1
libpng=1.6.37=h7cec526_2
libsodium=1.0.18=hbcb3906_1
libtiff=4.2.0=h355d032_0
libuv=1.40.0=h35c211d_0
libwebp-base=1.1.0=hbcb3906_3
llvm-openmp=11.0.0=h73239a0_1
lz4-c=1.9.3=h046ec9c_0
markupsafe=1.1.1=py39hcbf5805_3
matplotlib=3.3.3=py39h6e9494a_0
matplotlib-base=3.3.3=py39h5674cf4_0
mistune=0.8.4=py39hcbf5805_1003
mypy_extensions=0.4.3=py39h6e9494a_3
nbclassic=0.2.6=pyhd8ed1ab_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py39h6e9494a_3
nbformat=5.0.8=py_0
ncurses=6.2=h2e338ed_4
nest-asyncio=1.4.3=pyhd8ed1ab_0
nodejs=15.3.0=heaeed3d_0
notebook=6.1.6=py39h6e9494a_0
numpy=1.19.5=py39he588a01_1
olefile=0.46=pyh9f0ad1d_1
openssl=1.1.1i=h35c211d_0
packaging=20.8=pyhd3deb0d_0
pandas=1.2.0=py39h089d6f7_0
pandoc=2.11.3.2=h35c211d_0
pandocfilters=1.4.2=py_1
parso=0.8.1=pyhd8ed1ab_0
pathspec=0.8.1=pyhd3deb0d_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.1.0=py39hc06e9b5_1
pip=20.3.3=pyhd8ed1ab_0
prometheus_client=0.9.0=pyhd3deb0d_0
prompt-toolkit=3.0.10=pyha770c72_0
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.3=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.17.3=py39hcbf5805_2
pysocks=1.7.1=py39h6e9494a_3
python=3.9.1=h1d169a7_3_cpython
python-dateutil=2.8.1=py_0
python_abi=3.9=1_cp39
pytz=2020.5=pyhd8ed1ab_0
pyzmq=20.0.0=py39hd6373f1_1
readline=8.0=h0678c8f_2
regex=2020.11.13=py39hcbf5805_1
requests=2.25.1=pyhd3deb0d_0
scikit-learn=0.24.0=py39h1f17f7f_0
scipy=1.6.0=py39hca3e1c3_0
send2trash=1.5.0=py_0
setuptools=49.6.0=py39h6e9494a_3
six=1.15.0=pyh9f0ad1d_0
sniffio=1.2.0=py39h6e9494a_1
sqlite=3.34.0=h17101e1_0
terminado=0.9.2=py39h6e9494a_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=h0419947_1
toml=0.10.2=pyhd8ed1ab_0
tornado=6.1=py39hcbf5805_1
traitlets=5.0.5=py_0
typed-ast=1.4.2=py39hcbf5805_0
typing_extensions=3.7.4.3=py_0
tzdata=2020f=he74cb21_0
urllib3=1.26.2=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.36.2=pyhd3deb0d_0
xz=5.2.5=haf1e3a3_1
zeromq=4.3.3=h74dc148_3
zipp=3.4.0=py_0
zlib=1.2.11=h7795811_1010
zstd=1.4.8=hf387650_1
